+++ packages/snippets/src/browser/getContextInfo.js
@@ -3,7 +3,9 @@ const getElementXpath = require('./getElementXpath')
 function getContextInfo() {
   let isRoot
   try {
-    isRoot = window.top.document === window.document
+    isRoot =
+      window.top.document === window.document ||
+      window.document['applitools-marker'] === 'root-context'
   } catch (err) {
     isRoot = false
   }