+++ packages/snippets/CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Unreleased
 
+- adjust getContextInfo to work with Cypress Universal (return isRoot=true for context with document which is marked as "[applitools-marker]=root-context")
 
  ## 2.2.1 - 2022/3/12